pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lighthouse"
version = "0.1.0"
description = "Local/offline photo indexing and semantic search (no rearranging)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.109",
  "uvicorn>=0.25",
  "jinja2>=3.1",
  "pillow>=10.0",
  "platformdirs>=4.0",
  "typer>=0.9",
  "watchdog>=3.0",
]

[project.optional-dependencies]
minimal = []
clip = [
  "numpy>=1.24",
  "hnswlib>=0.8.0",
  "torch>=2.0",
  "open_clip_torch>=2.24.0",
]

[project.scripts]
lighthouse = "lighthouse.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
lighthouse = ["templates/*.html", "static/*.css", "static/*.js"]
